requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.6.0
djangorestframework==3.11.0
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
